+++ packages/terra-i18n/CHANGELOG.md
@@ -3,6 +3,22 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
+
+
+### Changed
+* Update tests for dev-site v6
+
+4.11.0 - (July 30, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+4.10.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 4.9.0 - (July 16, 2019)
 ------------------